Fix: change ollama stream to vercel stream Refactor: modify issue create
